+++ packages/import/i18n/en.js
@@ -1,0 +1,8 @@
+
+i18n.map('en', {
+	label_import: "Import Data",
+	label_importfile: "Upload a GeoJSON File",
+	label_importname: "Import Name",
+	label_imported: "Imported places",
+	error_import_formatNotValid: "Format not Valid, max ",
+});