+++ packages/import/i18n/fr.js
@@ -1,0 +1,5 @@
+
+i18n.map('en', {
+	label_import: "Import Data",
+	error_import_formatNotValid: "Format not Valid"
+});