+++ packages/import/i18n/it.js
@@ -1,0 +1,8 @@
+
+i18n.map('it', {
+	label_import: "Importa Dati",
+	label_importfile: "Carica un file GeoJSON",
+	label_importname: "Nome Importazione",	
+	label_imported: "Luoghi Importati",
+	error_import_formatNotValid: "Formato non Valido, max "
+});