+++ packages/import/plugin.js
@@ -1,0 +1,39 @@
+
+K.Plugin({
+	name: 'import',
+	templates: {
+		panelAdmin: 'panelAdmin_import',
+		panelPlace: 'panelPlace_import'
+	},
+	filters: {
+		placePanel: {
+			fields: {
+				import: 1
+			}
+		},
+		placeItem: {
+			fields: {
+				import: 1
+			}	
+		}
+	},
+	schemas: {
+		place: {
+			import: {}
+		}
+	},	
+	settings: {
+		"public": {
+			"import": {
+				"fileFormats": {
+					"application/json": true,
+					"application/geo+json": true
+				},
+				"maxFileSize": 2e+7 //20MB
+			}
+		},
+		/*//TODO "import": {
+			
+		}*/
+	}
+});