+++ CHANGELOG.md
@@ -4,6 +4,11 @@ Complete list of releases [here](https://github.com/Keplerjs/Kepler/releases)
 ## Next version
 ...
 
+## v1.4.8 - Modularity and Extensibility is ready
+- new plugin Import
+- new template placeholder panelAdmin
+- geoinfo plugin new settings autoupdate
+
 ## v1.4.7 - Modularity and Extensibility is ready
 - published all plugins in atmospherejs.com
 - remove workarounnd globsync from all plugins